Move DomainParticipant node out
