+++ dds/src/implementation/dds/mod.rs
@@ -1,1 +1,2 @@
 pub mod nodes;
+pub mod dds_domain_participant;
