+++ dds/src/implementation/dds/nodes.rs
@@ -14,22 +14,7 @@ use crate::{
     subscription::data_reader::AnyDataReader,
 };
 
-#[derive(Clone, PartialEq, Eq)]
-pub struct DomainParticipantNode {
-    participant_address: ActorAddress<DomainParticipantActor>,
-}
 
-impl DomainParticipantNode {
-    pub fn new(participant_address: ActorAddress<DomainParticipantActor>) -> Self {
-        Self {
-            participant_address,
-        }
-    }
-
-    pub fn participant_address(&self) -> &ActorAddress<DomainParticipantActor> {
-        &self.participant_address
-    }
-}
 
 #[derive(Clone, PartialEq, Eq)]
 pub struct SubscriberNode {
